receiver/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-bigquery >= 2.7.0
